[Refactor] 추가 job 수정
